requirements.txt
aiosmtplib==2.0.1
alembic==1.9.2
amqp==5.1.1
anyio==3.6.2
attrs==22.2.0
bcrypt==4.0.1
beautifulsoup4==4.11.2
billiard==3.6.4.0
black==22.12.0
blinker==1.5
bs4==0.0.1
celery==5.2.7
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
dnspython==2.3.0
ecdsa==0.18.0
email-validator==1.3.1
exceptiongroup==1.1.0

fastapi-jwt-auth==0.5.0
fastapi-mail==1.2.4
fastapi-paginate==0.1.0
fastapi-pagination==0.11.3
greenlet==2.0.1
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
idna==3.4
iniconfig==2.0.0
Jinja2==3.1.2
kombu==5.2.4
Mako==1.2.4
MarkupSafe==2.1.2
mypy-extensions==0.4.3
packaging==23.0
passlib==1.7.4
pathspec==0.10.3
platformdirs==2.6.2
pluggy==1.0.0
prompt-toolkit==3.0.36
py-dotenv==0.1
pyasn1==0.4.8
pydantic==1.10.4
PyJWT==1.7.1
pytest==7.2.1
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1

starlette==0.22.0
teamapt-monnify==1.1.1
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.14
uvicorn==0.20.0
vine==5.0.0
watchfiles==0.18.1
wcwidth==0.2.6
websockets==10.4
